requirements.txt
cmake
dlib
wheel
opencv-python
opencv-contrib-python
face_recognition
caer
selenium
Pillow
requests